requirements.txt
requests
pandas
openpyxl
msal
pyyaml
